danger evaluation #70
not tested yet, just the code

+++ src/control_unit/control_unit_A/logs.cpp
@@ -172,7 +172,7 @@ int checkFileFormat(char* newConfiguration) {
   } else {
     inZone = true;
 
-    float az, el;
+    double az, el;
     if (sscanf(line, "%f %f", &az, &el) != 2) {
       return -1;
     }
